requirements.txt
langgraph>=0.2.0
langchain>=0.2.0
langchain-core>=0.2.0
langchain-openai>=0.1.0
langchain-anthropic>=0.1.0
langchain-google-genai>=1.0.0
langchain-groq>=0.1.0
python-dotenv>=1.0.0
click>=8.1.0
rich>=13.0.0
pydantic>=2.0.0
pathlib2>=2.3.7
colorama>=0.4.6
